+++ client/src/features/game/Canvas/Canvas.tsx
@@ -1,25 +1,8 @@
-import React, { useState, useEffect, useRef, useCallback } from 'react';
+import React, { useEffect, useRef, useCallback } from 'react';
 import { ReactSketchCanvas, CanvasPath } from 'react-sketch-canvas';
 import { Box } from '@chakra-ui/react';
 
-enum StrokeWidth {
-  S = 2,
-  M = 4,
-  L = 12,
-  XL = 30,
-}
-
-// TODO: 색상 HEX값 지정
-enum StrokeColor {
-  Black = 'black',
-  Red = 'red',
-  Blue = 'blue',
-  Green = 'green',
-  Yellow = 'yellow',
-}
-
-const storkeWidths = Object.values(StrokeWidth);
-const strokeColors = Object.values(StrokeColor);
+import useCanvasToolbar from './useCanvasToolbar';
 
 interface CanvasProps {
   isEditMode: boolean;
@@ -29,8 +12,9 @@ function Canvas({ isEditMode }: CanvasProps) {
   const wrapperRef = useRef<HTMLDivElement>(null);
   const canvasRef = useRef<ReactSketchCanvas>(null);
 
-  const [strokeColor, setStrokeColor] = useState(StrokeColor.Black);
-  const [strokeWidth, setStrokeWidth] = useState(StrokeWidth.L);
+  const { strokeColor, strokeWidth, renderCanvasToolbar } = useCanvasToolbar({
+    canvasRef,
+  });
 
   useEffect(
     function preventEditCanvas() {
@@ -62,22 +46,6 @@ function Canvas({ isEditMode }: CanvasProps) {
     console.log(...updatedPath);
   }, []);
 
-  const handleClickStrokeColorSelectButton = useCallback(
-    (color: StrokeColor) => () => {
-      setStrokeColor(color);
-      canvasRef.current?.eraseMode(false);
-    },
-    [],
-  );
-
-  const handleClickEraserButton = useCallback(() => {
-    canvasRef.current?.eraseMode(true);
-  }, []);
-
-  const handleClickClearButton = useCallback(() => {
-    canvasRef.current?.clearCanvas();
-  }, []);
-
   return (
     <Box>
       <Box ref={wrapperRef}>
@@ -91,22 +59,7 @@ function Canvas({ isEditMode }: CanvasProps) {
           withTimestamp
         />
       </Box>
-      {isEditMode && (
-        <Box>
-          {strokeColors.map((color) => (
-            <Box
-              key={color}
-              as="button"
-              bg={color}
-              width="24px"
-              height="24px"
-              onClick={handleClickStrokeColorSelectButton(color)}
-            />
-          ))}
-          <button onClick={handleClickEraserButton}>지우개</button>
-          <button onClick={handleClickClearButton}>다 지우기</button>
-        </Box>
-      )}
+      {isEditMode && renderCanvasToolbar()}
     </Box>
   );
 }
