+++ client/src/features/game/Canvas/index.ts
@@ -1,0 +1,3 @@
+import Canvas from './Canvas';
+
+export default Canvas;
